Add files via upload
This is the Flexbox program! Some documentation is included, though I didn't do a good job at it.
Uses DSU, Dijkstras and Kruskals to compute results. Large inputs may be innacurate, I have not thoroughly tested this program yet!

I plan to add a front-end to this over the next week or so following this hackathon.